refactor: calculate scores method
