+++ commscores/commscores.py
@@ -19,210 +19,26 @@ from modelseedpy.core.fbahelper import FBAHelper
 from modelseedpy.core.msgapfill import MSGapfill
 from modelseedpy.core.msminimalmedia import MSMinimalMedia
 from modelseedpy.core.msmodelutl import MSModelUtil
-from multiprocess import current_process
-from numpy import array, load, nan, ndarray, sort, unique, where
-from numpy.random import shuffle
 from optlang import Constraint, Objective, Variable
 
-# from math import prod
-
+from .logger import logger
 
 warnings.simplefilter("ignore", category=DeprecationWarning)
 
-import logging
-
-logging.basicConfig(filename="example.log", encoding="utf-8", level=logging.ERROR)
-logger = logging.getLogger(__name__)
-
 
 rm_comp = FBAHelper.remove_compartment
-rm_costless = re.compile("(\s\(.+\))")
 
-
-def remove_metadata(element):
-    try:
-        element = float(rm_costless.sub("", str(element)).replace("%", ""))
-    except:
-        pass
-    return element
-
-
-def convert_to_int(element):
-    try:
-        element = int(element)
-    except:
-        pass
-    return element
-
-
-package_dir = os.path.abspath(os.path.dirname(__file__))
-categories_dir = os.path.join(package_dir, "data", "categories")
-sugars, aminoacids = (
-    load(os.path.join(categories_dir, "sugars.npy")),
-    load(os.path.join(categories_dir, "aminoAcids.npy")),
+from .utils import (
+    _categorize_mets,
+    _compatibilize,
+    _get_media,
+    _load_models,
+    _process_mets,
+    _sigfig_check,
+    convert_to_int,
+    nanFilter,
+    remove_metadata,
 )
-vitamins, minerals = (
-    load(os.path.join(categories_dir, "vitamins.npy")),
-    load(os.path.join(categories_dir, "minerals.npy")),
-)
-energy_compounds = load(os.path.join(categories_dir, "energy_compounds.npy"))
-sugars_dic, aminoacids_dic = (
-    dict(zip(sugars[:, 0], sugars[:, 1])),
-    dict(zip(aminoacids[:, 0], aminoacids[:, 1])),
-)
-vitamins_dic, minerals_dic = (
-    dict(zip(vitamins[:, 0], vitamins[:, 1])),
-    dict(zip(minerals[:, 0], minerals[:, 1])),
-)
-energy_compounds_dic = dict(zip(energy_compounds[:, 0], energy_compounds[:, 1]))
-
-
-def _categorize_mets(metIDs):
-    met_sugars, met_aminoAcids, met_vitamins, met_minerals, met_energy, met_other = (
-        [],
-        [],
-        [],
-        [],
-        [],
-        [],
-    )
-    for metID in metIDs:
-        if metID in sugars[:, 0]:
-            met_sugars.append(f"{sugars_dic[metID]} ({metID})")
-        elif metID in aminoacids[:, 0]:
-            met_aminoAcids.append(f"{aminoacids_dic[metID]} ({metID})")
-        elif metID in vitamins[:, 0]:
-            met_vitamins.append(f"{vitamins_dic[metID]} ({metID})")
-        elif metID in minerals[:, 0]:
-            met_minerals.append(f"{minerals_dic[metID]} ({metID})")
-        elif metID in energy_compounds[:, 0]:
-            met_energy.append(f"{energy_compounds_dic[metID]} ({metID})")
-        else:
-            met_other.append(metID)
-    return met_sugars, met_aminoAcids, met_vitamins, met_minerals, met_energy, met_other
-
-
-def _process_mets(metIDs):
-    return [", ".join(lst) for lst in _categorize_mets(metIDs)]
-
-
-def _compatibilize(member_models: Iterable, printing=False):
-    return member_models
-
-
-def _load_models(
-    member_models: Iterable, com_model=None, compatibilize=True, printing=False
-):
-    # ic(member_models, com_model, compatibilize)
-    if not com_model and member_models:
-        model = build_from_species_models(member_models, name="SMETANA_pair")
-        return member_models, model  # (model, names=names, abundances=abundances)
-    # models = PARSING_FUNCTION(community_model)  # TODO the individual models of a community model can be parsed
-    if compatibilize:
-        return (
-            _compatibilize(member_models, printing),
-            _compatibilize([com_model], printing)[0],
-        )
-    return member_models, com_model
-
-
-def _get_media(
-    media=None,
-    com_model=None,
-    model_s_=None,
-    min_growth=None,
-    environment=None,
-    interacting=True,
-    printing=False,
-    minimization_method="minFlux",
-    skip_bad_media=False,
-):
-    # print(media, com_model, model_s_)
-    if com_model is None and model_s_ is None:
-        raise TypeError("< com_model > or < model_s_ > must be parameterized.")
-    if media is not None:
-        if model_s_ is not None and not isinstance(model_s_, (list, set, tuple)):
-            return media["members"][model_s_.id]["media"]
-        elif com_model is not None:
-            return media["community_media"]
-        return media
-    # model_s_ is either a singular model or a list of models
-    if com_model is not None:
-        try:
-            com_media, media_sol = MSMinimalMedia.determine_min_media(
-                com_model,
-                minimization_method,
-                min_growth,
-                None,
-                interacting,
-                5,
-                printing,
-            )
-        except Exception as e:
-            if skip_bad_media:
-                com_media, media_sol = None, None
-            else:
-                print(e)
-    if model_s_ is not None:
-        if not isinstance(model_s_, (list, set, tuple, ndarray)):
-            try:
-                return MSMinimalMedia.determine_min_media(
-                    model_s_,
-                    minimization_method,
-                    min_growth,
-                    environment,
-                    interacting,
-                    printing,
-                )
-            except Exception as e:
-                if not skip_bad_media:
-                    print(e)
-                return None
-        members_media = {}
-        for model in model_s_:
-            try:
-                members_media[model.id] = {
-                    "media": MSMinimalMedia.determine_min_media(
-                        model,
-                        minimization_method,
-                        min_growth,
-                        environment,
-                        interacting,
-                        printing,
-                    )[0]
-                }
-                continue
-            except Exception as e:
-                if skip_bad_media:
-                    continue
-                else:
-                    print(e)
-        # print(members_media)
-        if com_model is None:
-            return members_media
-    else:
-        return com_media, media_sol
-    return {"community_media": com_media, "members": members_media}
-
-
-def _sigfig_check(value, sigfigs, default):
-    if str(value) in ["inf", "nan"]:
-        value = ""
-    if FBAHelper.isnumber(value):
-        return sigfig.round(value, sigfigs)
-    else:
-        return default
-
-
-def nanFilter(value, string=True):
-    if isinstance(value, str) or value is None:
-        if string:
-            return value
-        else:
-            return nan
-    if any([value < 0, value > 1e5]):
-        return "" if string else nan
-    return value
 
 
 class CommScores:
@@ -570,483 +386,6 @@ class CommScores:
             util.add_medium(environ)
             obj_vals.append(util.model.slim_optimize())
         return obj_vals
-
-    @staticmethod
-    def calculate_scores(
-        pairs,
-        models_media=None,
-        environments=None,
-        annotated_genomes=True,
-        lazy_load=False,
-        kbase_obj=None,
-        cip_score=True,
-        costless=True,
-        skip_bad_media=False,
-        check_models=True,
-        print_progress=False,
-    ):
-        from pandas import Series
-
-        if isinstance(pairs, list):
-            (
-                pairs,
-                models_media,
-                environments,
-                annotated_genomes,
-                lazy_load,
-                kbase_obj,
-            ) = pairs
-        series, mets = [], []
-        if not isinstance(environments, (list, tuple)):
-            environments = [environments]
-        if isinstance(environments, (list, tuple)) and hasattr(environments[0], "name"):
-            environments = {
-                m.name: FBAHelper.convert_kbase_media(m, 1000) for m in environments
-            }
-        elif not isinstance(environments, dict):
-            environments = {f"media{i}": m for i, m in enumerate(environments)}
-        pid = current_process().name
-        model_utils = {}
-        count = 0
-        for model1, models in pairs.items():
-            if model1.id == "":
-                model1.id = "model1"
-            if lazy_load:
-                model1, model1_str = CommScores._load(model1, kbase_obj)
-            else:
-                model1_str = model1.id
-            if model1.id not in models_media:
-                models_media[model1.id] = {
-                    "media": _get_media(model_s_=model1, skip_bad_media=skip_bad_media)
-                }
-                if models_media[model1.id] is None:
-                    continue
-            if model1.id not in model_utils:
-                model_utils[model1.id] = MSModelUtil(model1, True)
-            # print(pid, model1)
-            for model_index, model2 in enumerate(models):
-                if model2.id == "":
-                    model2.id = "model2"
-                if lazy_load:
-                    model2, model2_str = CommScores._load(model2, kbase_obj)
-                else:
-                    model2_str = model2.id
-                if model2.id not in models_media:
-                    models_media[model2.id] = {
-                        "media": _get_media(
-                            model_s_=model2, skip_bad_media=skip_bad_media
-                        )
-                    }
-                    if models_media[model2.id] is None:
-                        continue
-                if model2.id not in model_utils:
-                    model_utils[model2.id] = MSModelUtil(model2, True)
-                grouping = [model1, model2]
-                grouping_utils = [model_utils[model1.id], model_utils[model2.id]]
-                modelIDs = [model.id for model in grouping]
-                comm_model = build_from_species_models(grouping)
-                community = MSCommunity(comm_model, ids=modelIDs)
-                comm_sol = comm_model.optimize()
-                print(f"{pid}~~{count}\t{modelIDs}\t{comm_sol.objective_value}")
-                for environName, environ in environments.items():
-                    if print_progress:
-                        print(f"\tEnvironment\t{environName}", end="\t")
-                    if check_models:  # check that the models grow in the environment
-                        CommScores._check_model(
-                            model_utils[model1.id], environ, model1_str, skip_bad_media
-                        )
-                        CommScores._check_model(
-                            model_utils[model2.id], environ, model2_str, skip_bad_media
-                        )
-                    # initiate the KBase output
-                    report_dic = {
-                        f"model{i+1}": modelID for i, modelID in enumerate(modelIDs)
-                    }
-                    # the model growths are determined and the environmental media is parameterized for each of the members
-                    g1, g2, comm = [
-                        _sigfig_check(val, 5, "")
-                        for val in CommScores._determine_growths(
-                            [
-                                model_utils[model1.id],
-                                model_utils[model2.id],
-                                community.util,
-                            ],
-                            environ,
-                        )
-                    ]
-                    coculture_growths = {
-                        memID: abundance * comm
-                        for memID, abundance in community.predict_abundances().items()
-                    }
-                    report_dic.update(
-                        {
-                            "media": environName,
-                            "monoculture growth model1": g1,
-                            "monoculture growth model2": g2,
-                        }
-                    )
-                    report_dic.update(
-                        {
-                            f"coculture growth model{modelIDs.index(memID)+1}": growth
-                            for memID, growth in coculture_growths.items()
-                        }
-                    )
-                    report_dic.update({"community growth": comm})
-                    # define the MRO content
-                    mro_values = CommScores.mro(
-                        grouping, models_media, raw_content=True, environment=environ
-                    )
-                    report_dic.update(
-                        {
-                            f"MRO_model{modelIDs.index(models_string.split('--')[0])+1}": f"{100*len(intersection)/len(memMedia):.3f}% ({len(intersection)}/{len(memMedia)})"
-                            for models_string, (
-                                intersection,
-                                memMedia,
-                            ) in mro_values.items()
-                        }
-                    )
-                    mets.append({"MRO metabolites": list(mro_values.values())[0][0]})
-                    if print_progress:
-                        print("MRO done", end="\t")
-                    # define the CIP content
-                    if cip_score:
-                        cip_values = CommScores.cip(
-                            modelutils=[model_utils[mem.id] for mem in grouping]
-                        )
-                        report_dic.update({"CIP": cip_values[1]})
-                        mets[-1].update({"CIP metabolites": list(cip_values[0])})
-                        if print_progress:
-                            print("CIP done", end="\t")
-                    # define the MIP content
-                    mip_values = CommScores.mip(
-                        grouping,
-                        comm_model,
-                        0.1,
-                        None,
-                        None,
-                        environ,
-                        print_progress,
-                        True,
-                        costless,
-                        costless,
-                        skip_bad_media,
-                    )
-                    # print(mip_values)
-                    if mip_values is not None:
-                        report_dic.update(
-                            {
-                                f"MIP_model{modelIDs.index(models_name)+1}": str(
-                                    len(received)
-                                )
-                                for models_name, received in mip_values[0].items()
-                            }
-                        )
-                        mets[-1].update(
-                            {
-                                "MIP model1 metabolites": list(mip_values[0].values())[
-                                    0
-                                ],
-                                "MIP model2 metabolites": list(mip_values[0].values())[
-                                    1
-                                ],
-                            }
-                        )
-                        if costless:
-                            for models_name, received in mip_values[1].items():
-                                report_dic[
-                                    f"MIP_model{modelIDs.index(models_name)+1} (costless)"
-                                ] = (
-                                    report_dic[
-                                        f"MIP_model{modelIDs.index(models_name)+1}"
-                                    ]
-                                    + f" ({len(received)})"
-                                )
-                                del report_dic[
-                                    f"MIP_model{modelIDs.index(models_name)+1}"
-                                ]
-                            if print_progress:
-                                print("costless_MIP  done", end="\t")
-                    else:
-                        report_dic.update(
-                            {"MIP_model1 (costless)": "", "MIP_model2 (costless)": ""}
-                        )
-                        mets[-1].update(
-                            {
-                                "MIP model1 metabolites": [None],
-                                "MIP model2 metabolites": [None],
-                            }
-                        )
-                    if print_progress:
-                        print("MIP done", end="\t")
-                    # define the BSS content
-                    bss_values = CommScores.bss(
-                        grouping,
-                        grouping_utils,
-                        environments,
-                        models_media,
-                        skip_bad_media,
-                    )
-                    report_dic.update(
-                        {
-                            f"BSS_model{modelIDs.index(name.split(' supporting ')[0])+1}": f"{_sigfig_check(100*val, 5, '')}%"
-                            for name, (mets, val) in bss_values.items()
-                        }
-                    )
-                    mets[-1].update(
-                        {
-                            "BSS model1 metabolites": [
-                                met_set for met_set, val in bss_values.values()
-                            ][0],
-                            "BSS model2 metabolites": [
-                                met_set for met_set, val in bss_values.values()
-                            ][1],
-                        }
-                    )
-                    # mets[-1].update({"bss_mets": list(bss_values[0].values())})
-                    if print_progress:
-                        print("BSS done", end="\t")
-                    # define the PC content
-                    pc_values = CommScores.pc(
-                        grouping,
-                        grouping_utils,
-                        comm_model,
-                        None,
-                        comm_sol,
-                        environ,
-                        True,
-                        community,
-                    )
-                    report_dic.update(
-                        {
-                            "PC_model1": _sigfig_check(
-                                list(pc_values[1].values())[0], 5, ""
-                            ),
-                            "PC_model2": _sigfig_check(
-                                list(pc_values[1].values())[1], 5, ""
-                            ),
-                            "PC_comm": _sigfig_check(pc_values[0], 5, ""),
-                            "BIT": pc_values[3],
-                        }
-                    )
-                    if print_progress:
-                        print("PC  done\tBIT done", end="\t")
-                    # print([mem.slim_optimize() for mem in grouping])
-                    # define the GYD content
-                    logger.debug(
-                        list(
-                            CommScores.gyd(
-                                grouping,
-                                grouping_utils,
-                                environ,
-                                False,
-                                community,
-                                check_models,
-                            ).values()
-                        )
-                    )
-                    gyd1, gyd2, g1, g2 = list(
-                        CommScores.gyd(
-                            grouping,
-                            grouping_utils,
-                            environ,
-                            False,
-                            community,
-                            check_models,
-                        ).values()
-                    )[0]
-                    report_dic.update(
-                        {
-                            "GYD1": _sigfig_check(gyd1, 5, ""),
-                            "GYD2": _sigfig_check(gyd2, 5, ""),
-                        }
-                    )
-                    if print_progress:
-                        print("GYD done\t\t", end="\t" if annotated_genomes else "\n")
-                    # define the FS content
-
-                    if kbase_obj is not None and annotated_genomes:
-                        fs_values = list(
-                            CommScores.fs(
-                                grouping, kbase_obj, annotated_genomes=annotated_genomes
-                            ).values()
-                        )[0]
-                        print(
-                            len(fs_values[0]) if fs_values[0] is not None else "NaN",
-                            fs_values[1],
-                        )
-                        report_dic.update({"FS": sigfig.round(fs_values[1], 5)})
-                        if fs_values is not None:
-                            mets[-1].update({"FS features": fs_values[0]})
-                        if print_progress:
-                            print("FS done\t\t")
-                    # return a pandas Series, which can be easily aggregated with other results into a DataFrame
-                    series.append(Series(report_dic))
-                count += 1
-        return series, mets
-
-    @staticmethod
-    def report_generation(
-        all_models: iter = None,  # a list of distinct lists is provided for specifying exclusive groups
-        pairs: dict = None,
-        mem_media: dict = None,
-        pair_limit: int = None,
-        exclude_pairs: list = None,
-        kbase_obj=None,
-        annotated_genomes: dict = True,  # True triggers internal acquisition of the genomes, where None skips
-        see_media=True,
-        environments: iter = None,  # a collection of environment dicts or KBase media objects
-        pool_size: int = None,
-        cip_score=True,
-        costless=True,
-        skip_bad_media=False,
-        check_models=True,
-        print_progress=False,
-    ):
-        from pandas import concat
-
-        if pairs:
-            model_pairs = unique(
-                [
-                    {model1, model2}
-                    for model1, models in pairs.items()
-                    for model2 in models
-                ]
-            )
-        elif all_models is not None:
-            if not isinstance(all_models[0], list):
-                all_models = list(set(all_models))
-                model_pairs = array(list(combinations(all_models, 2)))
-            else:
-                model_pairs = []
-                for models1, models2 in combinations(all_models, 2):
-                    models1 = set(models1)
-                    models2 = set(models2)
-                    if len(models1) > len(models2):
-                        larger_list = models1
-                        smaller_list = models2
-                    else:
-                        larger_list = models2
-                        smaller_list = models1
-                    model_pairs.append(
-                        [
-                            list(zip(combin, smaller_list))
-                            for combin in permutations(larger_list, len(smaller_list))
-                        ]
-                    )
-                # flatten the assembled pairs and filter duplicates
-                model_pairs = array(
-                    [
-                        x
-                        for x in set(
-                            tuple(x)
-                            for x in [
-                                i
-                                for y in list(chain.from_iterable(model_pairs))
-                                for i in y
-                            ]
-                        )
-                    ]
-                )
-                all_models = list(chain.from_iterable(all_models))
-            if pair_limit is not None:
-                shuffle(model_pairs)
-                new_pairs = []
-                for index, pair in enumerate(model_pairs):
-                    if set(pair) not in exclude_pairs and index < pair_limit:
-                        new_pairs.append(pair)
-                    elif index >= pair_limit:
-                        break
-                model_pairs = array(new_pairs)
-            if isinstance(model_pairs[0], str):
-                model_pairs = unique(sort(model_pairs, axis=1))
-            pairs = {
-                first: model_pairs[where(model_pairs[:, 0] == first)][:, 1]
-                for first in model_pairs[:, 0]
-            }
-        else:
-            raise ValueError(
-                "Either < all_models > or < pairs > must be defined to simulate interactions."
-            )
-        if not all_models:
-            all_models = list(
-                chain(*[list(values) for values in pairs.values()])
-            ) + list(pairs.keys())
-        lazy_load = len(model_pairs) > 10000  # all_models[0], (list,set,tuple))
-        if lazy_load and not kbase_obj:
-            ValueError(
-                "The < kbase_obj > argument must be provided to lazy load models."
-            )
-        new_models = []
-        for index, model in enumerate(all_models):
-            if model.id == "":
-                model.id = f"model_index{index}"
-            new_models.append(model)
-        all_models = new_models[:]
-        if not mem_media:
-            models_media = _get_media(
-                model_s_=all_models, skip_bad_media=skip_bad_media
-            )
-        else:
-            models_media = mem_media.copy()
-            missing_models = set()
-            missing_modelID = []
-            for model in all_models:
-                if model is not None and model.id not in models_media:
-                    missing_models.add(model)
-                    missing_modelID.append(
-                        model if not hasattr(model, "id") else model.id
-                    )
-            if missing_models != set():
-                logger.error(
-                    f"Media of the {missing_modelID} models are not defined, and will be calculated separately."
-                )
-                models_media.update(
-                    _get_media(model_s_=missing_models), skip_bad_media=skip_bad_media
-                )
-        if see_media:
-            print(f"The minimal media of all members:\n{models_media}")
-        print(f"\nExamining the {len(list(model_pairs))} model pairs")
-        if pool_size is not None:
-            from datetime import datetime
-
-            from multiprocess import Pool
-
-            print(
-                f"Loading {int(pool_size)} workers and computing the scores",
-                datetime.now(),
-            )
-            pool = Pool(
-                int(pool_size)
-            )  # .map(calculate_scores, [{k: v} for k,v in pairs.items()])
-            args = [
-                [
-                    dict([pair]),
-                    models_media,
-                    environments,
-                    annotated_genomes,
-                    lazy_load,
-                    kbase_obj,
-                ]
-                for pair in list(pairs.items())
-            ]
-            output = pool.map(CommScores.calculate_scores, args)
-            series = chain.from_iterable([ele[0] for ele in output])
-            mets = chain.from_iterable([ele[1] for ele in output])
-        else:
-            series, mets = CommScores.calculate_scores(
-                pairs,
-                models_media,
-                environments,
-                annotated_genomes,
-                lazy_load,
-                kbase_obj,
-                cip_score,
-                costless,
-                skip_bad_media,
-                check_models,
-                print_progress,
-            )
-        return concat(series, axis=1).T, mets
 
     @staticmethod
     def mro(
